Refactor y mejora de la gestión de clientes
Se ha cambiado el tipo de parámetro del método `Delete` en `CustomerController.cs` de `long` a `int`. Se han añadido nuevos mapeos en `ConfigurationServices.cs` para `CustomerRequest` y `Customer`.

En `IRepository.cs`, el método `Remove` ha sido modificado para ser asíncrono y ahora devuelve un `Task`, además de añadir un nuevo método `RemoveRange`. En `ICustomerServices.cs`, el método `DeleteCustomer` también ha cambiado su parámetro de `long` a `int`.

En `CustomerServices.cs`, se ha implementado el método `CreateCustomer` con lógica de validación y manejo de excepciones. El método `DeleteCustomer` ahora utiliza `int` y se ha añadido lógica para eliminar un cliente y sus publicaciones asociadas, junto con métodos auxiliares para facilitar estas operaciones.

En `Customer.cs`, se ha añadido una propiedad `Posts` para representar la relación entre `Customer` y `Post`. En `HttpResponse.cs`, se han añadido métodos estáticos `Success` y `Fail` para simplificar la creación de respuestas HTTP. Finalmente, en `Repository.cs`, el método `Remove` ahora marca la entidad como eliminada en lugar de eliminarla físicamente, y se ha implementado `RemoveRange` para manejar múltiples eliminaciones.

+++ API/Controllers/CustomerController.cs
@@ -38,7 +38,7 @@ namespace API.Controllers.Customer
         }
 
         [HttpDelete("DeleteCustomer/{Id}")]
-        public async Task<IActionResult> Delete(long Id)
+        public async Task<IActionResult> Delete(int Id)
         {
             var result = await _customerServices.DeleteCustomer(Id);
             return StatusCode((int)result.HttpStatusCode, result);
